+++ CHANGES.rst
@@ -5,7 +5,11 @@ Changelog of nens-auth-client
 0.11 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Extend SSOMigrationBackend so that it matches external users by username
+  (from email) if it is an Azure AD account ending with @nelen-schuurmans.nl.
+
+- Dropped support for Python 3.5. Added Python 3.9 to the versions to be
+  tested.
 
 
 0.10 (2021-02-23)